Replace callcc with break
I believe this is semantically equivalent but it could use some more
eyes before it's ready for the main trunk.  This is the only use of
callcc within the framework and getting rid of it has other advantages
aside from fixing a crash -- namely, non-yarv ruby implementations tend
not to implement continuations at all, so this gets us closer to
supporting jruby and rbx.

See #4991